+++ packages/web-components/src/components/global-header/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.62.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.61.0...@carbon-labs/wc-global-header@0.62.0) (2026-04-15)
+
+
+### Features
+
+* **global-header:** cookie preferences ([#1178](https://github.com/carbon-design-system/carbon-labs/issues/1178)) ([9c3ef00](https://github.com/carbon-design-system/carbon-labs/commit/9c3ef00febe178e6f024c87c7ca1958a8e5515ae))
+
+
+
+
+
+# [0.61.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.60.0...@carbon-labs/wc-global-header@0.61.0) (2026-04-14)
+
+**Note:** Version bump only for package @carbon-labs/wc-global-header
+
+
+
+
+
 # [0.60.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.59.0...@carbon-labs/wc-global-header@0.60.0) (2026-04-09)
 
 **Note:** Version bump only for package @carbon-labs/wc-global-header